E_CLASSROOM-253 Added new component file
