+++ src/pages/admin/quizzes/QuizEdit/index.js
@@ -1,20 +1,13 @@
-import React, { Fragment, useState } from 'react';
+import React, { Fragment } from 'react';
 import { Link } from 'react-router-dom';
 import Container from 'react-bootstrap/Container';
 import Button from 'react-bootstrap/Button';
 import Nav from 'react-bootstrap/Nav';
-import Form from 'react-bootstrap/Form';
-import Dropdown from 'react-bootstrap/Dropdown';
-import { RiArrowDropDownLine } from 'react-icons/ri';
-import MultipleChoiceType from './components/MultipleChoiceType';
-import IdentificationType from './components/IdentificationType';
+import QuestionType from './components/QuestionType';
 
 import style from './index.module.scss';
 
 const QuizEdit = () => {
-  const [questionType, setQuestionType] = useState('multiple_choice');
-  const [timeLimit, setTimeLimit] = useState(0);
-
   const quiz = {
     id: 1,
     title: 'Web Development Basics',
@@ -38,49 +31,6 @@ const QuizEdit = () => {
       question_type: 'identification',
     },
   ];
-
-  const types = [
-    {
-      name: 'Multiple Choice',
-      value: 'multiple_choice',
-    },
-    {
-      name: 'Identification',
-      value: 'identification',
-    },
-  ];
-
-  const limit = ['5', '10', '20', '30', '60'];
-
-  const question = (e) => {
-    setQuestionType(e);
-  };
-
-  const time = (e) => {
-    setTimeLimit(e);
-  };
-
-  const question_type = (choice) => {
-    switch (choice) {
-    case 'multiple_choice':
-      return <MultipleChoiceType questions={questions[0]} />;
-    case 'identification':
-      return <IdentificationType questions={questions[1]} />;
-    default:
-      break;
-    }
-  };
-
-  const choice = (option) => {
-    switch (option) {
-    case 'multiple_choice':
-      return 'Multiple Choice';
-    case 'identification':
-      return 'Identification';
-    default:
-      break;
-    }
-  };
 
   return (
     <div className="d-inline-flex">
@@ -107,61 +57,7 @@ const QuizEdit = () => {
             <Button className={style.sidebarButtons}>Add a Question</Button>
             <Button className={style.sidebarButtons}>Change Category</Button>
           </div>
-          <div>{question_type(questionType)}</div>
-          <div className={style.formGap}>
-            <Dropdown onSelect={question}>
-              <Form>
-                <Form.Label className={style.inputTitle}>
-                  Question Type
-                </Form.Label>
-              </Form>
-              <Dropdown.Toggle
-                variant="link"
-                id="dropdown-basic"
-                bsPrefix="none"
-                className={style.dropdownStyle}
-              >
-                {choice(questionType)}
-                <RiArrowDropDownLine className={style.iconSize} />
-              </Dropdown.Toggle>
-              <Dropdown.Menu className={style.dropdownMenuStyle}>
-                {types.map((type, idx) => {
-                  return (
-                    <Dropdown.Item key={idx} eventKey={type.value}>
-                      {type.name}
-                    </Dropdown.Item>
-                  );
-                })}
-              </Dropdown.Menu>
-            </Dropdown>
-            <div className={style.formSpacing}>
-              <Dropdown onSelect={time}>
-                <Form>
-                  <Form.Label className={style.inputTitle}>
-                    Time Limit
-                  </Form.Label>
-                </Form>
-                <Dropdown.Toggle
-                  variant="link"
-                  id="dropdown-basic"
-                  bsPrefix="none"
-                  className={style.dropdownStyle}
-                >
-                  {timeLimit} seconds
-                  <RiArrowDropDownLine className={style.iconSize} />
-                </Dropdown.Toggle>
-                <Dropdown.Menu className={style.dropdownMenuStyle}>
-                  {limit.map((time, idx) => {
-                    return (
-                      <Dropdown.Item key={idx} eventKey={time}>
-                        {time} seconds
-                      </Dropdown.Item>
-                    );
-                  })}
-                </Dropdown.Menu>
-              </Dropdown>
-            </div>
-          </div>
+          <QuestionType questions={questions} />
         </div>
         <div className={style.confirmationButtons}>
           <Link to={`/admin/quizzes/${quiz.id}`} className={style.cancelButton}>
